+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 Add addition operators on Dictionaries.
+Add int subscript to String extension.
 
 ## [Unreleased](https://github.com/openium/SwiftiumKit/compare/latest...HEAD)
 ### Added