+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Add addition operators on Dictionaries.
 Add int subscript to String extension.
 Add closed int range subscript to String extension.
+Add int range subscript to String extension.
 
 ## [Unreleased](https://github.com/openium/SwiftiumKit/compare/latest...HEAD)
 ### Added