+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add int range subscript to String extension.
 - Add negative int subscript to String extension.
 - Add int subscript setter to String extension.
+- Add firstLowercased utility to String extension.
 
 ### Changed